Rename tests and files to remove opencl where not necessary, clarify the docs
