+++ stan/math/gpu/copy_triangular_transposed.hpp
@@ -1,5 +1,5 @@
-#ifndef STAN_MATH_GPU_COPY_TRIANGULAR_TRANSPOSED_OPENCL_HPP
-#define STAN_MATH_GPU_COPY_TRIANGULAR_TRANSPOSED_OPENCL_HPP
+#ifndef STAN_MATH_GPU_COPY_TRIANGULAR_TRANSPOSED_HPP
+#define STAN_MATH_GPU_COPY_TRIANGULAR_TRANSPOSED_HPP
 #ifdef STAN_OPENCL
 #include <stan/math/gpu/matrix_gpu.hpp>
 #include <stan/math/gpu/err/check_square.hpp>
