+++ stan/math/gpu/kernels/add_matrix_kernel.cl
@@ -11,9 +11,9 @@ R"(
 /**
  * Matrix addition on the GPU
  *
- * @param C (Write) Output matrix.
- * @param A (Read) LHS of matrix addition.
- * @param B (Read) RHS of matrix addition.
+ * @param[out] C Output matrix.
+ * @param[in] A LHS of matrix addition.
+ * @param[in] B RHS of matrix addition.
  * @param rows Number of rows for matrix A.
  * @param cols Number of rows for matrix B.
  *
