+++ test/unit/math/gpu/copy_submatrix_test.cpp
@@ -1,6 +1,6 @@
 #ifdef STAN_OPENCL
 #include <stan/math/prim/mat.hpp>
-#include <stan/math/gpu/copy_submatrix_opencl.hpp>
+#include <stan/math/gpu/copy_submatrix.hpp>
 #include <gtest/gtest.h>
 #include <algorithm>
