+++ test/unit/math/gpu/identity_test.cpp
@@ -1,6 +1,6 @@
 #ifdef STAN_OPENCL
 #include <stan/math/prim/mat.hpp>
-#include <stan/math/gpu/identity_opencl.hpp>
+#include <stan/math/gpu/identity.hpp>
 #include <gtest/gtest.h>
 #include <algorithm>
