+++ test/unit/math/gpu/zeros_test.cpp
@@ -1,6 +1,6 @@
 #ifdef STAN_OPENCL
 #include <stan/math/prim/mat.hpp>
-#include <stan/math/gpu/zeros_opencl.hpp>
+#include <stan/math/gpu/zeros.hpp>
 #include <gtest/gtest.h>
 #include <algorithm>
